requirements.txt
yt-dlp
python-mpv
requests
py7zr
